[ refactor ] SettingContent 元件拆分
